refactor: 코드 리뷰 반영

+++ src/main/java/me/jincrates/pf/assignment/application/dto/UpdateProductRequest.java
@@ -3,7 +3,6 @@ package me.jincrates.pf.assignment.application.dto;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import jakarta.validation.constraints.Positive;
 import jakarta.validation.constraints.PositiveOrZero;
-import java.math.BigDecimal;
 import java.util.Set;
 import lombok.With;
 
@@ -13,9 +12,9 @@ public record UpdateProductRequest(
     Long id,
     String name,
     @Positive(message = "상품 판매가는 0원보다 커야합니다.")
-    BigDecimal sellingPrice,
+    Long sellingPrice,
     @PositiveOrZero(message = "상품 할인가는 0원 이상이여야 합니다.")
-    BigDecimal discountPrice,
+    Long discountPrice,
     String brand,
     Set<Long> categoryIds
 ) {
